rm padding on amd
